refactor: reorganize font handling and improve font loading mechanism
