+++ system/ui/lib/text_measure.py
@@ -1,5 +1,6 @@
 import pyray as rl
-from openpilot.system.ui.lib.application import FONT_SCALE, font_fallback
+from openpilot.system.ui.lib.application import gui_app
+from openpilot.system.ui.lib.font import FONT_SCALE, font_fallback
 from openpilot.system.ui.lib.emoji import find_emoji
 
 _cache: dict[int, rl.Vector2] = {}
@@ -7,7 +8,7 @@ _cache: dict[int, rl.Vector2] = {}
 
 def measure_text_cached(font: rl.Font, text: str, font_size: int, spacing: int = 0) -> rl.Vector2:
   """Caches text measurements to avoid redundant calculations."""
-  font = font_fallback(font)
+  font = font_fallback(font, gui_app)
   key = hash((font.texture.id, text, font_size, spacing))
   if key in _cache:
     return _cache[key]
