+++ system/ui/lib/wrap_text.py
@@ -1,4 +1,5 @@
 import pyray as rl
+from openpilot.system.ui.lib.application import gui_app
 from openpilot.system.ui.lib.text_measure import measure_text_cached
 from openpilot.system.ui.lib.application import font_fallback
 
@@ -41,7 +42,7 @@ _cache: dict[int, list[str]] = {}
 
 
 def wrap_text(font: rl.Font, text: str, font_size: int, max_width: int) -> list[str]:
-  font = font_fallback(font)
+  font = font_fallback(font, gui_app)
   key = hash((font.texture.id, text, font_size, max_width))
   if key in _cache:
     return _cache[key]
